requirements.txt
groq==0.9.0
websockets==12.0
python-dotenv
aiohttp